requirements.txt
asgiref==3.8.1
blinker==1.8.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==5.0.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.2
django-import-export==4.0.3
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
et_xmlfile==2.0.0
Flask==3.0.3
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openpyxl==3.1.5
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
reportlab==4.4.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
uritemplate==4.1.1
Werkzeug==3.0.3
